+++ storage_utils/utils.py
@@ -1,7 +1,13 @@
 from contextlib import contextmanager
+import logging
 import os
+import shlex
 from subprocess import check_call
+from swiftclient.service import SwiftUploadObject
 import tempfile
+
+
+logger = logging.getLogger(__name__)
 
 
 def is_swift_path(p):
@@ -67,48 +73,111 @@ def path(p):
         return PosixPath(p)
 
 
-def copy(source, dest, copy_cmd='cp -r', object_threads=20,
-         segment_threads=20, **retry_args):
-    """Copies a source path to a destination path. Assumes that paths are
-    capable of being copied to/from.
+def copy(source, dest, swift_retry_options=None):
+    """Copies a source file to a destination file.
+
+    Note that this utility can be called from either swift or posix
+    paths created with `storage_utils.path`.
 
     Args:
-        src (path|str): The source path to copy from
-        dest (path|str): The destination to copy to
-        copy_cmd (str): If copying to / from posix, this command is
-            used.
-        object_threads (int): The amount of object threads to use
-            for swift uploads / downloads.
-        segment_threads (int): The amount of segment threads to use
-            for swift uploads.
-        retry_args (dict): Optional retry arguments to use for swift upload
-            or download. View the swift module-level documentation for more
+        source (path|str): The source directory to copy from
+        dest (path|str): The destination file. In contrast to
+            ``shutil.copy``, the parent directory is created if it doesn't
+            already exist.
+        swift_retry_options (dict): Optional retry arguments to use for swift
+            upload or download. View the
+            `swift module-level documentation <swiftretry>` for more
             information on retry arguments
+
+    Examples:
+        Copying a swift file to a local path behaves as follows::
+
+            >>> from storage_utils import path
+            >>> swift_p = path('swift://tenant/container/dir/file.txt')
+            >>> # file.txt will be copied to other_dir/other_file.txt
+            >>> swift_p.copy('other_dir/other_file.txt')
+
+        Copying from a local path to swift behaves as follows::
+
+            >>> from storage_utils import path
+            >>> local_p = path('my/local/file.txt')
+            >>> # File will be uploaded to swift://tenant/container/dir/my_file.txt
+            >>> local_p.copy('swift://tenant/container/dir/')
+
+        Because of the ambiguity in whether a remote target is a file or directory, copy()
+        will error on ambiguous paths.
+
+            >>> local_p.copy('swift://tenant/container/dir')
+            Traceback (most recent call last):
+            ...
+            ValueError: swift destination must be file with extension or directory with slash
     """
     source = path(source)
     dest = path(dest)
+    swift_retry_options = swift_retry_options or {}
     if is_swift_path(source) and is_swift_path(dest):
-        raise ValueError('Cannot copy one swift path to another swift path')
+        raise ValueError('cannot copy one swift path to another swift path')
+    if is_swift_path(dest) and dest.is_ambiguous():
+        raise ValueError('swift destination must be file with extension or directory with slash')
 
-    if not is_swift_path(dest):
-        # Ensure the parent directory exists on the destination for cp or mcp
-        # to run properly
-        dest.expand().abspath().parent.makedirs_p()
-
-    if is_swift_path(source):
-        source.download(output_dir=dest, remove_prefix=True,
-                        object_threads=object_threads)
-    elif is_swift_path(dest):
-        with source:
-            dest.upload(['.'],
-                        segment_size=1073741824,
-                        use_slo=True,
-                        object_threads=object_threads,
-                        segment_threads=segment_threads)
+    dest_file = dest if dest.name else dest / source.name
+    if is_posix_path(dest):
+        dest_file.makedirs_p()
+        if is_swift_path(source):
+            source._download_object(dest_file, **swift_retry_options)
+        else:
+            copy_cmd = ['cp',
+                        str(source.abspath().expand()),
+                        str(dest.abspath().expand())]
+            logger.info('performing copy with command - %s', copy_cmd)
+            check_call(copy_cmd)
     else:
-        formatted_copy_cmd = copy_cmd.split()
-        formatted_copy_cmd.extend([str(source), str(dest)])
-        check_call(formatted_copy_cmd)
+        if not dest_file.parent.container:
+            raise ValueError((
+                'cannot copy to tenant "%s" and file '
+                '"%s"' % (dest_file.parent, dest_file.name)
+            ))
+        dest_obj_name = path(dest_file.parent.resource or '') / dest_file.name
+        dest_file.parent.upload([SwiftUploadObject(source, object_name=dest_obj_name)],
+                                **swift_retry_options)
+        return dest_file.parent, source
+
+
+def copytree(source, dest, copy_cmd='cp -r', swift_upload_options=None,
+             swift_download_options=None):
+    """Copies a source directory to a destination directory. Assumes that
+    paths are capable of being copied to/from.
+
+    Args:
+        source (path|str): The source directory to copy from
+        dest (path|str): The directory to copy to
+        copy_cmd (str): If copying to / from posix, this command is
+            used.
+        swift_upload_options (dict): When the destination is a swift path,
+            pass these options as keyword arguments to `SwiftPath.upload`.
+        swift_download_options (dict): When the source is a swift path,
+            pass these options as keyword arguments to `SwiftPath.download`.
+    """
+    source = path(source)
+    dest = path(dest)
+    swift_upload_options = swift_upload_options or {}
+    swift_download_options = swift_download_options or {}
+    if is_swift_path(source) and is_swift_path(dest):
+        raise ValueError('cannot copy one swift path to another swift path')
+
+    if is_posix_path(dest):
+        dest.expand().abspath().parent.makedirs_p()
+        if is_swift_path(source):
+            source.download(dest, **swift_download_options)
+        else:
+            copy_cmd = shlex.split(copy_cmd)
+            copy_cmd.extend([str(source.abspath().expand()),
+                             str(dest.abspath().expand())])
+            logger.info('performing copy with command - %s', copy_cmd)
+            check_call(copy_cmd)
+    else:
+        with source:
+            dest.upload(['.'], **swift_upload_options)
 
 
 def walk_files_and_dirs(files_and_dirs):